HU7 - Implementacion de autovalidacion para las solicitudes de prestamos que cumplan con los requisitos necesatios. En esta HU tambien se integran las lambdas que realizan los calculos pertinentes para determinar si un prestamo puede ser aprobado automaticamente y las que notifican al usuario por medio de SNS cuando cambia el estado del prestamo y en caso de que sea aprovado el plan de pagos mes a mes

+++ domain/model/src/main/java/com/crediya/model/loanrequest/DomainLoanRequestsDTO.java
@@ -1,6 +1,7 @@
 package com.crediya.model.loanrequest;
 
 import java.math.BigDecimal;
+import java.util.UUID;
 
 public record DomainLoanRequestsDTO(String email, //user
                                     String name, //user
@@ -8,7 +9,7 @@ public record DomainLoanRequestsDTO(String email, //user
                                     String loanType, //loanType
 
                                     BigDecimal amount, //loanReq
-                                    Double timeLimit, //loanReq
+                                    int timeLimit, //loanReq
 
                                     Double interestRate, //loanType
